+++ packages/orq-rc/RELEASES.md
@@ -4809,3 +4809,13 @@ Based on:
 - [typescript v4.0.0-rc.10] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.10 - packages/orq-rc
+
+## 2025-11-06 05:50:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.649.0 (2.739.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.11] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.11 - packages/orq-rc